add prometheus metric
