+++ api/package.json
@@ -15,6 +15,7 @@
     "cors": "^2.8.5",
     "dotenv": "16.3.1",
     "express": "4.17.1",
+    "prom-client": "^15.1.1",
     "socks-proxy-agent": "^8.0.1",
     "unique-names-generator": "^4.7.1"
   },
